fix coupla implicit type conversions in src/

+++ src/convert-osm-rcpp.cpp
@@ -355,7 +355,8 @@ void convert_multipoly_to_sp (Rcpp::S4 &multipolygons, const Relations &rels,
             polygons.slot ("Polygons") = outList_i;
             // convert id_vec to single string
             std::string id_vec_str = id_vec [i] [0];
-            for (int j = 1; j < id_vec [i].size (); j++)
+            for (int j = 1;
+                    j < static_cast <unsigned int> (id_vec [i].size ()); j++)
                 id_vec_str += "." + id_vec [i] [j];
             polygons.slot ("ID") = id_vec_str;
             //polygons.slot ("ID") = id_vec [i]; // sp expects char not vec!
@@ -375,7 +376,7 @@ void convert_multipoly_to_sp (Rcpp::S4 &multipolygons, const Relations &rels,
     // Fill plotOrder slot with int vector - this has to be int, not
     // unsigned int!
     std::vector <int> plotord (rels.size ());
-    for (int j=0; j<rels.size (); j++)
+    for (int j=0; j<static_cast <unsigned int> (rels.size ()); j++)
         plotord [j] = j + 1;
     multipolygons.slot ("plotOrder") = plotord;
     plotord.clear ();
